rename : 패키지 이름 partybymatch > partyinfobymatch 로 변경
