+++ src/main/java/com/playus/twpservice/domain/party/dto/partyinfobymatch/PartyInfoListByMatchRequest.java
@@ -1,4 +1,4 @@
-package com.playus.twpservice.domain.party.dto.partybymatch;
+package com.playus.twpservice.domain.party.dto.partyinfobymatch;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
